Effect 3 4

+++ 3DLv2_2022_vs2019_Students/GameProgramming/src/CEffect.h
@@ -25,7 +25,7 @@ public:
 	void Render();
 	void Render(CMaterial* material);
 private:
-	float mFadeCount;
+	int mFadeCount;
 	EType mType;
 	//s”
 	int mRows;
